+++ internal/db/migrations/postgres/schema.sql
@@ -91,6 +91,16 @@ CREATE INDEX bookmark_updated_idx ON "bookmark" USING btree (updated DESC);
 CREATE INDEX bookmark_url_idx ON "bookmark" (url);
 CREATE INDEX bookmark_initial_url_idx ON "bookmark" (initial_url);
 
+CREATE TABLE IF NOT EXISTS bookmark_removed (
+    uid     varchar(32) NOT NULL,
+    user_id integer     NOT NULL,
+    deleted timestamptz NOT NULL DEFAULT NOW(),
+
+    CONSTRAINT fk_bookmark_removed_user FOREIGN KEY (user_id) REFERENCES "user"(id) ON DELETE CASCADE
+);
+
+CREATE INDEX bookmark_removed_deleted_idx ON "bookmark_removed" (deleted DESC);
+
 --
 -- Search configuration
 --
@@ -151,6 +161,24 @@ $$;
 
 CREATE TRIGGER bookmark_tsu AFTER INSERT OR UPDATE ON bookmark
 	FOR EACH ROW EXECUTE PROCEDURE bookmark_search_update();
+
+CREATE OR REPLACE FUNCTION bookmark_after_delete()
+RETURNS trigger
+LANGUAGE plpgsql
+AS $$
+BEGIN
+    INSERT INTO bookmark_removed (
+        uid, user_id
+    ) VALUES (
+        OLD.uid, OLD.user_id
+    );
+
+    RETURN OLD;
+END;
+$$;
+
+CREATE TRIGGER bookmark_ad AFTER DELETE on bookmark
+    FOR EACH ROW EXECUTE PROCEDURE bookmark_after_delete();
 
 CREATE TABLE IF NOT EXISTS bookmark_collection (
     id          SERIAL      PRIMARY KEY,